Arreglo del codigo, separacion del batch y los itembatch en archivos distintos y correccion de las rutas correspondientes a los nuevos archivos
